(SERVER) refactor app structure
