+++ server/src/modules/post/post.routes.ts
@@ -1,5 +1,5 @@
 import express from 'express';
-import * as PostController from '../controllers/post.controller';
+import * as PostController from './post.controller';
 
 const router = express.Router();
